+++ frb_example/dart_minimal/flutter_rust_bridge.yaml
@@ -7,5 +7,5 @@ local: true
 
 # TODO temp
 #full_dep: true
-enable_lifetime: true
-stop_on_error: true
+#enable_lifetime: true
+#stop_on_error: true
