+++ frb_example/pure_dart/rust/src/api/misc_no_twin_example_a.rs
@@ -3,6 +3,7 @@
 use crate::frb_generated::RustAutoOpaque;
 use flutter_rust_bridge::frb;
 pub use std::any::Any;
+use std::sync::{Arc, Mutex};
 
 // Reproduce #1630
 #[frb(opaque)]
@@ -131,4 +132,24 @@ impl DeliberateFailSanityCheckTwinNormal {
 
 pub fn function_with_arg_type_name_override(a: Box<dyn Any + Send + Sync + 'static>) {
     let _ = a;
+}
+
+#[derive(Debug, Clone)]
+#[frb(opaque)]
+pub struct SimpleLogger(Arc<Mutex<Vec<String>>>);
+
+impl SimpleLogger {
+    #[frb(sync)]
+    pub fn new() -> Self {
+        Self(Arc::new(Mutex::new(vec![])))
+    }
+
+    pub(crate) fn log(&self, message: &str) {
+        self.0.lock().unwrap().push(message.to_owned());
+    }
+
+    #[frb(sync)]
+    pub fn get_and_reset(&self) -> Vec<String> {
+        self.0.lock().unwrap().drain(..).collect()
+    }
 }
